+++ tapo/src/api/hub_handler.rs
@@ -101,18 +101,43 @@ impl HubHandler {
     /// or to support all the possible devices connected to the hub.
     /// If the deserialization fails, or if a property that you care about it's not present, try [`HubHandler::get_child_device_list_json`].
     pub async fn get_child_device_list(&self) -> Result<Vec<ChildDeviceHubResult>, Error> {
-        self.client
-            .read()
-            .await
-            .get_child_device_list::<ChildDeviceListHubResult>()
-            .await
-            .map(|r| r.devices)
+        let mut results = Vec::new();
+        let mut start_index = 0;
+        let mut fetch = true;
+
+        while fetch {
+            let devices = self
+                .client
+                .read()
+                .await
+                .get_child_device_list::<ChildDeviceListHubResult>(start_index)
+                .await
+                .map(|r| r.devices)?;
+
+            fetch = devices.len() == 10;
+            start_index += 10;
+            results.extend(devices);
+        }
+
+        Ok(results)
     }
 
     /// Returns *child device list* as [`serde_json::Value`].
     /// It contains all the properties returned from the Tapo API.
-    pub async fn get_child_device_list_json(&self) -> Result<serde_json::Value, Error> {
-        self.client.read().await.get_child_device_list().await
+    ///
+    /// # Arguments
+    ///
+    /// * `start_index` - the index to start fetching the child device list.
+    ///   It should be `0` for the first page, `10` for the second, and so on.
+    pub async fn get_child_device_list_json(
+        &self,
+        start_index: u64,
+    ) -> Result<serde_json::Value, Error> {
+        self.client
+            .read()
+            .await
+            .get_child_device_list(start_index)
+            .await
     }
 
     /// Returns *child device component list* as [`serde_json::Value`].