+++ tapo/src/requests.rs
@@ -2,6 +2,7 @@
 
 mod control_child;
 mod energy_data_interval;
+mod get_child_device_list;
 mod get_energy_data;
 mod get_trigger_logs;
 mod handshake;
@@ -18,6 +19,7 @@ pub use play_alarm::*;
 pub use set_device_info::*;
 
 pub(crate) use control_child::*;
+pub(crate) use get_child_device_list::*;
 pub(crate) use get_energy_data::*;
 pub(crate) use get_trigger_logs::*;
 pub(crate) use handshake::*;